git-svn-id: svn+ssh://green.exnet.com/rw/SVN-repositories/public/reutils/trunk@234 7a6dc53a-678a-4e31-bf73-3c095f94da06

+++ javasrc/org/hd/d/edh/FUELINST.java
@@ -207,9 +207,11 @@ public final class FUELINST
 
         /**Total grid distribution and transmission losses in range [0.0,1.0]. */
         public final float totalGridLosses;
-        
-        /**Current/latest intensity correlation with fuel type, immutable; may be empty but not null. */
-        public final Map<String,Double> correlationIntensityToFuel;
+
+        /**Current/latest intensity correlation with fuel type, immutable; may be empty but not null.
+         * Likely actual precision only warrants float.
+         */
+        public final Map<String,Float> correlationIntensityToFuel;
 
 
         /**Construct a default (empty) instance.
@@ -220,7 +222,7 @@ public final class FUELINST
             this(null, null, 0L, 0L, 0L, 0, Collections.<String,Integer>emptyMap(), 0L, 0, 0L, 0, 0, 0L, 0L, 0, 0, 0,
                     null, null, null, null,
                     0f,
-                    Collections.<String,Double>emptyMap());
+                    Collections.<String,Float>emptyMap());
             }
 
         /**Construct an instance.
@@ -244,7 +246,7 @@ public final class FUELINST
                 final List<Integer> histAveZCGenerationByHourOfDay,
                 final List<Integer> histAveStorageDrawdownByHourOfDay,
                 final float totalGridLosses,
-                final Map<String,Double> correlationIntensityToFuel)
+                final Map<String,Float> correlationIntensityToFuel)
             {
             this.status = status;
             this.recentChange = recentChange;
@@ -271,7 +273,7 @@ public final class FUELINST
             this.histAveStorageDrawdownByHourOfDay = new SummaryByHour(histAveStorageDrawdownByHourOfDay); // Defensive copy.
             this.totalGridLosses = totalGridLosses;
             if(correlationIntensityToFuel == null) { throw new IllegalArgumentException(); }
-            this.correlationIntensityToFuel = Collections.unmodifiableMap(new HashMap<String, Double>(correlationIntensityToFuel)); // Defensive copy.
+            this.correlationIntensityToFuel = Collections.unmodifiableMap(new HashMap<String, Float>(correlationIntensityToFuel)); // Defensive copy.
             }
 
         /**Select the traffic-light colour for a given intensity (gCO2/kWh); never null.
